% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_model_prelim.R
\name{diagnose_model}
\alias{diagnose_model}
\title{Diagnose an Ordinal Model}
\usage{
diagnose_model(model)
}
\arguments{
\item{model}{An ordinal model object (e.g., from \code{polr} in the \code{MASS} package).}
}
\value{
A data frame containing diagnostic metrics for the model, including:
\item{model_iteration}{The iteration number of the model.}
\item{adjacent_predictions_count}{The number of predictions where the two most probable categories are adjacent.}
\item{mean_surrogate_res}{The mean of the surrogate residuals.}
\item{sd_surrogate_res}{The standard deviation of the surrogate residuals.}
\item{aic}{The Akaike Information Criterion of the model.}
\item{bic}{The Bayesian Information Criterion of the model.}
\item{ivs}{The independent variables in the model.}
\item{n_interactions}{The number of interaction terms in the model.}
\item{n_vars_model}{The number of variables in the model.}
}
\description{
This function takes an ordinal model and returns a data frame with various diagnostic metrics.
}
\examples{
\dontrun{
# Assuming ord_model is a fitted ordinal model
diagnostics <- diagnose_model(ord_model)
print(diagnostics)
}

}
