% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_post_demo.R
\name{graph_post_demo_diagnose}
\alias{graph_post_demo_diagnose}
\title{Create a Density Ridge Plot for Post-Demographic Diagnosis}
\usage{
graph_post_demo_diagnose(data)
}
\arguments{
\item{data}{A data frame containing the results of the post-demographic diagnosis. The data frame
should have columns `real_class`, `predicted_class`, and `predicted_probability`.}
}
\value{
A ggplot2 object representing the density ridge plot.
}
\description{
This function creates a density ridge plot for the results of post-demographic diagnosis,
showing the distribution of predicted probabilities across different classes.
}
\examples{
# Example usage:
data <- data.frame(
  real_class = factor(sample(c("totally_biden", "somewhat_biden", "undecided", "somewhat_trump", "totally_trump"), 100, replace = TRUE)),
  predicted_class = factor(sample(c("totally_biden", "somewhat_biden", "undecided", "somewhat_trump", "totally_trump"), 100, replace = TRUE)),
  predicted_probability = runif(100)
)
plot <- graph_post_demo_diagnose(data)
print(plot)

}
