% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_model_prelim.R
\name{is_adjacent}
\alias{is_adjacent}
\title{Check if the Top Two Predicted Categories are Adjacent}
\usage{
is_adjacent(probs)
}
\arguments{
\item{probs}{A numeric vector of probabilities for each category.}
}
\value{
A logical value indicating whether the two most probable categories are adjacent (TRUE) or not (FALSE).
}
\description{
This function checks if the two most probable categories (based on their probabilities) are adjacent in an ordinal scale.
}
\examples{
probs <- c(0.1, 0.2, 0.3, 0.25, 0.15)
is_adjacent(probs)

}
