% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_model_prelim.R
\name{is_extreme}
\alias{is_extreme}
\title{Check if the Top Two Predicted Categories are Extremes}
\usage{
is_extreme(probs)
}
\arguments{
\item{probs}{A numeric vector of probabilities for each category.}
}
\value{
A logical value indicating whether the two most probable categories are the two extremes (TRUE) or not (FALSE).
}
\description{
This function checks if the two most probable categories (based on their probabilities) are the two extremes in an ordinal scale.
}
\examples{
probs <- c(0.4, 0.1, 0.1, 0.1, 0.3)
is_extreme(probs)

}
