% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_model_prelim.R
\name{random_stan_polr}
\alias{random_stan_polr}
\title{Generate a Random Ordinal Regression Model using rstanarm::stan_polr}
\usage{
random_stan_polr(
  data,
  n_vars_app = 13,
  n_vars_model = 27,
  n_interactions,
  necessary_variables_prefixes = necessary_variables_prefixes,
  extra_variables_prefixes = extra_variables_prefixes,
  refresh = 0
)
}
\arguments{
\item{data}{The data frame containing the dataset.}

\item{n_vars_app}{The number of additional variables to sample for the model.}

\item{n_vars_model}{The total number of variables to include in the model.}

\item{n_interactions}{The number of interaction terms to include in the model.}

\item{necessary_variables_prefixes}{A vector of prefixes for the necessary variables.}

\item{extra_variables_prefixes}{A vector of prefixes for the extra variables.}
}
\value{
A fitted `stan_polr` model object with additional information on the number of interactions and variables used.
}
\description{
This function randomly selects variables and interaction terms to generate
an ordinal regression model using `rstanarm::stan_polr`.
}
\examples{
# Assuming `data` is your dataset and `necessary_variables_prefixes` and `extra_variables_prefixes` are defined:
random_model <- random_stan_polr(data, 13, 27, 5, necessary_variables_prefixes, extra_variables_prefixes)
}
