% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_post_demo.R
\name{update_and_predict_polr_model}
\alias{update_and_predict_polr_model}
\title{Update and Predict Proportional Odds Logistic Regression Model (polr)}
\usage{
update_and_predict_polr_model(model, new_train_data, test_data)
}
\arguments{
\item{model}{A proportional odds logistic regression model of class `polr` to be updated.}

\item{new_train_data}{A data frame containing the new training data to update the model.}

\item{test_data}{A data frame containing the test data for which predictions are to be made.}
}
\value{
A data frame containing the predictions with columns `id`, `predicted_class`, and `predicted_probability`.
}
\description{
This function updates a proportional odds logistic regression model (polr) with new training data and generates predictions on test data.
}
\examples{
# Example usage:
model <- MASS::polr(dv_vote ~ ivs_formula, data = train_data, Hess = TRUE)
new_train_data <- train_data_sample
test_data <- test_data_sample
predictions <- update_and_predict_polr_model(model, new_train_data, test_data)
print(predictions)

}
