% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_model_prelim.R
\name{create_model_prelim}
\alias{create_model_prelim}
\title{Create Preliminary Model}
\usage{
create_model_prelim(
  data,
  lm = FALSE,
  n_vars_app = 13,
  n_vars_model = 10,
  n_interactions,
  necessary_variables_prefixes = necessary_variables_prefixes,
  extra_variables_prefixes = extra_variables_prefixes
)
}
\arguments{
\item{data}{A data frame containing the dataset to be used for model creation.}

\item{n_vars_app}{An integer specifying the number of extra variables to use.}

\item{n_vars_model}{An integer specifying the number of interaction terms to generate.}

\item{necessary_variables_prefixes}{A vector of prefixes for necessary variables.}

\item{extra_variables_prefixes}{A vector of prefixes for extra variables.}
}
\value{
An object of class \code{polr} representing the fitted model.
}
\description{
This function creates a preliminary ordinal logistic regression model using
randomly selected extra variables and interaction terms. The function will
repeatedly attempt to create a model until successful.
}
\examples{
\dontrun{
data <- data.frame(vote = sample(c("totally Biden", "somewhat Biden", "neutral", "somewhat Trump", "totally Trump"), 100, replace = TRUE),
                   var1 = rnorm(100), var2 = rnorm(100), var3 = rnorm(100), var4 = rnorm(100))
necessary_variables_prefixes <- c("var1", "var2")
extra_variables_prefixes <- c("var3", "var4")
n_vars_app <- 2
n_vars_model <- 10
model <- create_model_prelim(data, n_vars_app, n_vars_model, necessary_variables_prefixes, extra_variables_prefixes)
}
}
